Fixed build on 64bits RHEL5.
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/epics/trunk/modules/motor@377 ececa60f-9f63-0410-97c7-ea05888eb260

+++ motorApp/ImsSrc/devIM483SM.cc
@@ -113,7 +113,7 @@ static struct board_stat **IM483SM_cards;
 STATIC long IM483SM_init(void *arg)
 {
     long rtnval;
-    int after = (int) arg;
+    long after = (long) arg;
 
     if (after == 0)
     {
